Update README.md
The word "tech report" should be "technical report" to make the sentence more grammatically correct. It should be:

"If you are using the pulsar backend for sphere-rendering (the PulsarPointRenderer or pytorch3d.renderer.points.pulsar.Renderer), please cite the technical report:"

This change ensures that the text is grammatically accurate.